+++ model/model_cash_flow.go
@@ -14,11 +14,11 @@ type CashFlow struct {
 	// Category of this cash flow. When null, then this is the cash flow of transactions that do not have a category.
 	Category *Category `json:"category,omitempty"`
 	// The total calculated income for the given category
-	Income float32 `json:"income"`
+	Income float64 `json:"income"`
 	// The total calculated spending for the given category
-	Spending float32 `json:"spending"`
+	Spending float64 `json:"spending"`
 	// The calculated balance for the given category
-	Balance float32 `json:"balance"`
+	Balance float64 `json:"balance"`
 	// The total count of income transactions for the given category
 	CountIncomeTransactions int32 `json:"countIncomeTransactions"`
 	// The total count of spending transactions for the given category
